Prototype OS
This is just kind of a basic way I had imagined for the interface.  I
don't know what to use as the screen right now.  I want to make the
buttons smaller as well as making them actually have arrows instead of
words for the VM.  We can talk about how to design the actual OS on
Tuesday.