Create Basic-Data.txt
This is a collection of basic biotopic data that can be used for a basic database.  Ideally, the specific type of tank can be selected from one of these options, and then installed modules will be able to subsequently pull data such as ranges and limits for the purposes of security, validation, etc.